Universal API for Swift and ObjC
The main goal of this commit is to defer allocation/computation work associated
with a log statement until the last possible moment, including doing none of the work
at all if the current log level does not support the log request.

• Swift: Use autoclosure so caller's strings are not generated until needed
• ObjC: Use macros aided by two new public API in AspenHelpers.swift

Tests for both new and old behavior have also been added.